Fix mknod dev
