+++ test-wrapper/src/lib/permissions.rs
@@ -29,8 +29,8 @@ impl StatBase for libc::stat {
         self.st_gid = group;
     }
 
-    fn set_rdev(&mut self, dev: dev_t) {
-        self.st_rdev = dev;
+    fn set_rdev(&mut self, rdev: dev_t) {
+        self.st_rdev = rdev;
     }
 }
 
@@ -52,8 +52,8 @@ impl StatBase for libc::stat64 {
         self.st_gid = group;
     }
 
-    fn set_rdev(&mut self, dev: dev_t) {
-        self.st_rdev = dev;
+    fn set_rdev(&mut self, rdev: dev_t) {
+        self.st_rdev = rdev;
     }
 }
 
@@ -74,10 +74,9 @@ fn stat_base(path: CPath, statbuf: &mut StatBase) {
     if let Some(group) = overrides.group {
         statbuf.set_group(group);
     }
-    // TODO this breaks stuff
-    //if let Some(dev) = overrides.dev {
-    //    statbuf.set_rdev(dev);
-    //}
+    if let Some(rdev) = overrides.rdev {
+        statbuf.set_rdev(rdev);
+    }
 }
 
 fn chmod_base<'a, F: Fn(mode_t) -> c_int>(path: CPath, mode: mode_t, orig_chmod: F) -> Result<c_int> {
@@ -101,20 +100,6 @@ fn chown_base(path: CPath, owner: uid_t, group: gid_t) -> Result<c_int> {
     let group = if (group as i32) == -1 { None } else { Some(group) };
     message(Message::OverrideOwner(path.get_id()?, owner, group))?;
     Ok(0)
-}
-
-fn mknod_base<'a, F: Fn() -> CPath, M: Fn(mode_t) -> c_int>(get_path: F, mode: mode_t, dev: dev_t, mknod: M) -> Result<c_int> {
-    let override_mode = mode & libc::S_IFCHR == libc::S_IFCHR || mode & libc::S_IFBLK == libc::S_IFBLK;
-    let base_mode = if override_mode {
-        libc::S_IFREG | 0o600 | (mode & 0o777)
-    } else {
-        mode
-    };
-    let ret = mknod(base_mode);
-    if ret == 0 && override_mode {
-        let _ = message(Message::OverrideMode(get_path().get_id()?, mode, mode_t::max_value(), Some(dev)));
-    }
-    Ok(ret)
 }
 
 wrap! {
@@ -179,14 +164,6 @@ wrap! {
             stat_base(CPath::from_path_at(dfd, path, flags), &mut *statbuf);
         }
         Ok(ret)
-    }
-
-    unsafe fn mknod:ORIG_MKNOD(path: *const c_char, mode: mode_t, dev: dev_t) -> c_int {
-        mknod_base(|| CPath::from_path(path, false), mode, dev, |mode| ORIG_MKNOD(path, mode, dev))
-    }
-
-    unsafe fn mknodat:ORIG_MKNODAT(dfd: c_int, path: *const c_char, mode: mode_t, dev: dev_t) -> c_int {
-        mknod_base(|| CPath::from_path_at(dfd, path, 0), mode, dev, |mode| ORIG_MKNODAT(dfd, path, mode, dev))
     }
 }
 
@@ -254,14 +231,6 @@ wrap! {
             stat_base(CPath::from_path_at(dfd, path, flags), &mut *statbuf);
         }
         Ok(ret)
-    }
-
-    unsafe fn __xmknod:ORIG_XMKNOD(ver: c_int, path: *const c_char, mode: mode_t, dev: dev_t) -> c_int {
-        mknod_base(|| CPath::from_path(path, false), mode, dev, |mode| ORIG_XMKNOD(ver, path, mode, dev))
-    }
-
-    unsafe fn __xmknodat:ORIG_XMKNODAT(ver: c_int, dfd: c_int, path: *const c_char, mode: mode_t, dev: dev_t) -> c_int {
-        mknod_base(|| CPath::from_path_at(dfd, path, 0), mode, dev, |mode| ORIG_XMKNODAT(ver, dfd, path, mode, dev))
     }
 }
